+++ actionpack/CHANGELOG.md
@@ -1,6 +1,11 @@
+*   Routing: Refactor `:action` default handling to ensure that path
+    parameters are not mutated during route generation.
+
+    *Andrew White*
+
 *   Add extension synonyms `yml` and `yaml` for MIME type `application/x-yaml`.
 
-   *bogdanvlviv*
+    *bogdanvlviv*
 
 *   Adds support for including ActionController::Cookies in API controllers.
     Previously, including the module would raise when trying to define